Edit what forms appear on landing and on clicks
When a user lands on the page, show the  sign in and sign up forms.
When the user signs in, hide the sign in and sign up forms, but show
the game board, sign out, and change password forms.
When the user signs out, show the sign in/up forms, and hide the rest.

Probably not the best way to do it, but will fix later if I have time.
Did not break anything.