Add create game functionality
When the user clicks "create game" button, a new game is added to the
user in the API (yay!). The games numbers continue to accumulate and
do not reset after sign out.

Note to self: I am not creating the data layout, I am only USING
what was already created by the API. I only need to know
how to access the data within them/update data within the structure.