Add game update functionality
Game update PATCH functionality works!
When the user clicks on a box, the value in the box is updated in the
proper index of the API game[cell] array.

Next step is to make sure when user clicks "Create Game," the
current game board erases. And I need to fix up the code so it's not
a bunch of console.log's and the user actually gets responses.
Also need to update "over" functionality??