Add game stats, clear input text, hide passwords
Get Games functionality is hidden until the user is signed in.
When get games button is clicked, user receives message with number
of games played.
When a user enters credentials/form information, on success, the data
within the forms is cleared out.
When the user enters passwords, they are comcealed.

Removed console.logs