Make pretty and debug
Original plan was to just complete the CSS, which is did.
Upon double checking, I realized my clicks weren't pushing to the
cells array on update. It was passing a blank value. After much
troubleshooting, I realized there was an element out of order that
pushed the value to the array before the setClickValue method.

Not sure when that changed from last night, but it's fixed now.

Need to update one last css change, then double test, then remove all
debugging console.logs.