Fix broken code
Issue - if a user signs out mid game, or if a user clicks the create-
new game button in the middle of a game, they get a new game, but
the values don't clear, so the user gets error messages that boxes have
already been filled in.

Will continue working on it.